Função validate_book movida para o arquivo utils.py
